+++ tests/test_cli.py
@@ -44,8 +44,6 @@ def test_simpleaf_with_map_runs(tmp_path):
         str(input_path),
         "--output",
         str(output_path),
-        "--chemistry",
-        "10X_3p_v3",
         "--skip_umap_tsne",
         "--verbose",
     ]
@@ -55,65 +53,6 @@ def test_simpleaf_with_map_runs(tmp_path):
     print(result.stderr)
     assert result.returncode == 0
     assert output_path.exists()
-
-
-def test_simpleaf_with_cb_list(tmp_path):
-    input_path = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc/quants.h5ad")
-    output_path = tmp_path / "test_output_cookie" / "simpleaf_latest_1k_with_cb_list"
-    valid_cb_list = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc/total_retained_cells.tsv")
-    main_py = Path(__file__).resolve().parent.parent / "src" / "qcatch" / "main.py"
-
-    env = os.environ.copy()
-    env["PYTHONPATH"] = str(Path(__file__).resolve().parent.parent / "src")
-
-    cmd = [
-        "python",
-        str(main_py),
-        "--input",
-        str(input_path),
-        "--output",
-        str(output_path),
-        "--chemistry",
-        "10X_3p_v3",
-        "--valid_cell_list",
-        str(valid_cb_list),
-        "--skip_umap_tsne",
-        "--export_summary_table",
-        "--verbose",
-    ]
-
-    result = subprocess.run(cmd, capture_output=True, text=True, env=env)
-    print(result.stdout)
-    print(result.stderr)
-    assert result.returncode == 0
-    assert output_path.exists()
-
-
-def test_simpleaf_overwrite_save(tmp_path):
-    input_path = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc_overwrite/quants.h5ad")
-    valid_cb_list = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc/total_retained_cells.tsv")
-    main_py = Path(__file__).resolve().parent.parent / "src" / "qcatch" / "main.py"
-
-    env = os.environ.copy()
-    env["PYTHONPATH"] = str(Path(__file__).resolve().parent.parent / "src")
-
-    cmd = [
-        "python",
-        str(main_py),
-        "--input",
-        str(input_path),
-        "--chemistry",
-        "10X_3p_v3",
-        "--valid_cell_list",
-        str(valid_cb_list),
-        "--skip_umap_tsne",
-        "--verbose",
-    ]
-
-    result = subprocess.run(cmd, capture_output=True, text=True, env=env)
-    print(result.stdout)
-    print(result.stderr)
-    assert result.returncode == 0
 
 
 def test_mtx_input_1k_pbmc_v3(tmp_path):
@@ -141,3 +80,63 @@ def test_mtx_input_1k_pbmc_v3(tmp_path):
     print(result.stderr)
     assert result.returncode == 0
     assert output_path.exists()
+
+
+# def test_simpleaf_with_cb_list(tmp_path):
+#     input_path = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc/quants.h5ad")
+#     output_path = tmp_path / "test_output_cookie" / "simpleaf_latest_1k_with_cb_list"
+#     valid_cb_list = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc/total_retained_cells.tsv")
+#     main_py = Path(__file__).resolve().parent.parent / "src" / "qcatch" / "main.py"
+
+#     env = os.environ.copy()
+#     env["PYTHONPATH"] = str(Path(__file__).resolve().parent.parent / "src")
+
+#     cmd = [
+#         "python",
+#         str(main_py),
+#         "--input",
+#         str(input_path),
+#         "--output",
+#         str(output_path),
+#         "--chemistry",
+#         "10X_3p_v3",
+#         "--valid_cell_list",
+#         str(valid_cb_list),
+#         "--skip_umap_tsne",
+#         "--export_summary_table",
+#         "--verbose",
+#     ]
+
+#     result = subprocess.run(cmd, capture_output=True, text=True, env=env)
+#     print(result.stdout)
+#     print(result.stderr)
+#     assert result.returncode == 0
+#     assert output_path.exists()
+
+
+# def test_simpleaf_overwrite_save(tmp_path):
+#     input_path = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc_overwrite/quants.h5ad")
+#     valid_cb_list = Path("tests/data/test_data/simpleaf_rerun_1k_pbmc/total_retained_cells.tsv")
+#     main_py = Path(__file__).resolve().parent.parent / "src" / "qcatch" / "main.py"
+
+#     env = os.environ.copy()
+#     env["PYTHONPATH"] = str(Path(__file__).resolve().parent.parent / "src")
+
+#     cmd = [
+#         "python",
+#         str(main_py),
+#         "--input",
+#         str(input_path),
+#         "--chemistry",
+#         "10X_3p_v3",
+#         "--valid_cell_list",
+#         str(valid_cb_list),
+#         "--skip_umap_tsne",
+#         "--export_summary_table",
+#         "--verbose",
+#     ]
+
+#     result = subprocess.run(cmd, capture_output=True, text=True, env=env)
+#     print(result.stdout)
+#     print(result.stderr)
+#     assert result.returncode == 0