+++ CHANGELOG.md
@@ -8,6 +8,19 @@ and this project adheres to [Semantic Versioning][].
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.2.7] 2025-11-04
+### Added
+- Support python 3.13.
+### Changed
+
+- Added automatic inference of the chemistry version from simpleaf quantification metadata.
+- Removed the default chemistry assumption — QCatch now requires either:
+    1. a successfully inferred chemistry from simpleaf’s metadata,
+    2. an explicitly specified chemistry via --chemistry (-c), or
+    3. a custom `number of partitions` provided via --n_partitions (-n).
+
+    If none of these are supplied, QCatch will stop and prompt the user to specify one.
+
 ## [0.2.6] 2025-06-29
 
 ### Added